Homework14 completed

+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = this.max();
-        for (int sale : sales) {
+    public long min() {
+        long min = this.max();
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -25,9 +25,9 @@ public class SalesManager {
         return min;
     }
 
-    public int cropAverage() {
-        int sum = 0;
-        for (int sale : sales) {
+    public long cropAverage() {
+        long sum = 0;
+        for (long sale : sales) {
             sum += sale;
         }
         return (sum - min() - max()) / (sales.length - 2);
